ScreenManager takes a list of screens.
This list will be used when a new screen is created. So instead of
having to access the screen outside of the screen manager directly, we
reference a screen by the string with which it was stored in the screen
list.

This prevents issues when several screens need to interact with each
other. Now they only have to know the name of the screen they want to
push onto the stack.